[fix] 연진이랑 하던 PurchaseActivity merge conflict 해결
